# backend/docs/swagger.yaml
definitions:
...
  database.Passenger:
    properties:
      booking_id:
        type: integer
      citizen_id:
        $ref: '#/definitions/sql.NullString'
      dob:
        type: string
      email:
        $ref: '#/definitions/sql.NullString'
      first_name:
        type: string
      gender:
        type: string
      last_name:
        type: string
      passenger_id:
        type: integer
      phone_number:
        $ref: '#/definitions/sql.NullString'
      seat_number:
        type: string
    type: object
  database.User:
    properties:
      citizen_id:
        type: string
      dob:
        type: string
      email:
        $ref: '#/definitions/sql.NullString'
      first_name:
        type: string
      gender:
        type: string
      hashed_password:
        type: string
      last_name:
        type: string
      phone_number:
        $ref: '#/definitions/sql.NullString'
      user_id:
        type: integer
      username:
        type: string
    type: object
  sql.NullString:
    properties:
      string:
        type: string
      valid:
        description: Valid is true if String is not NULL
        type: boolean
    type: object
info:
  contact: {}
paths:
...
  /api/v1/passengers:
    get:
      consumes:
      - application/json
      description: Get all passengers
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Passenger'
            type: array
      summary: Get all passengers
      tags:
      - passengers
    post:
      consumes:
      - application/json
      description: Create a passenger
      parameters:
      - description: Passenger object
        in: body
        name: passenger
        required: true
        schema:
          $ref: '#/definitions/database.Passenger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Passenger'
      summary: Create a passenger
      tags:
      - passengers
  /api/v1/passengers/booking/{booking_id}:
    get:
      consumes:
      - application/json
      description: Get all passengers by booking ID
      parameters:
      - description: Booking ID
        in: path
        name: booking_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.Passenger'
            type: array
      summary: Get all passengers by booking ID
      tags:
      - passengers
  /api/v1/passengers/passenger/{passenger_id}:
    delete:
      consumes:
      - application/json
      description: Remove a passenger
      parameters:
      - description: Passenger ID
        in: path
        name: passenger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Remove a passenger
      tags:
      - passengers
    get:
      consumes:
      - application/json
      description: Get passenger by ID
      parameters:
      - description: Passenger ID
        in: path
        name: passenger_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Passenger'
      summary: Get passenger by ID
      tags:
      - passengers
    put:
      consumes:
      - application/json
      description: Update passenger by ID
      parameters:
      - description: Passenger ID
        in: path
        name: passenger_id
        required: true
        type: integer
      - description: Passenger object
        in: body
        name: passenger
        required: true
        schema:
          $ref: '#/definitions/database.Passenger'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.Passenger'
      summary: Update passenger by ID
      tags:
      - passengers
  /api/v1/users:
    get:
      consumes:
      - application/json
      description: Get all users
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/database.User'
            type: array
      summary: Get all users
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create a user
      parameters:
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/database.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.User'
      summary: Create a user
      tags:
      - users
  /api/v1/users/search/citizen_id/{citizen_id}:
    get:
      consumes:
      - application/json
      description: Get user by citizen ID
      parameters:
      - description: Citizen ID
        in: path
        name: citizen_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.User'
      summary: Get user by citizen ID
      tags:
      - users
  /api/v1/users/user/{user_id}:
    delete:
      consumes:
      - application/json
      description: Remove user by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.User'
      summary: Remove user by ID
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Get user by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.User'
      summary: Get user by ID
      tags:
      - users
    put:
      consumes:
      - application/json
      description: Update user by ID
      parameters:
      - description: User ID
        in: path
        name: user_id
        required: true
        type: integer
      - description: User object
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/database.User'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/database.User'
      summary: Update user by ID
      tags:
      - users
swagger: "2.0"